requirements.txt
# -------------------------
# Core
# -------------------------
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
gunicorn

# -------------------------
# Authentication & Security
# -------------------------
Flask-Login
Flask-JWT-Extended
Flask-WTF
Flask-Limiter
Flask-Caching

# -------------------------
# Serialization / Validation
# -------------------------
marshmallow
Flask-Marshmallow

# -------------------------
# Background Tasks
# -------------------------
celery
redis

# -------------------------
# Email & Admin
# -------------------------
Flask-Mail
Flask-Admin
Flask-Babel

# -------------------------
# Database Driver
# -------------------------
psycopg2-binary

# -------------------------
# Testing
# -------------------------
pytest
pytest-cov
faker

# -------------------------
# Code Quality
# -------------------------
# black
ruff